Move FileSystem completion utilities to the editor features layer
At the moment, these aren't used by any completion providers in the features layer. Until we decide that this is ready to support cross-platform, let's keep them in the editor features layer closer to the providers that actually use them.
