+++ src/EditorFeatures/CSharp/CSharpEditorFeatures.csproj
@@ -94,13 +94,13 @@
     <Compile Include="ChangeSignature\ChangeSignatureCommandHandler.cs" />
     <Compile Include="Classification\CSharpEditorClassificationService.cs" />
     <Compile Include="CommentSelection\CSharpCommentUncommentService.cs" />
-    <Compile Include="Completion\CompletionProviders\GlobalAssemblyCacheCompletionHelper.cs" />
     <Compile Include="Completion\CompletionProviders\OverrideCompletionProvider.cs" />
     <Compile Include="Completion\CompletionProviders\PartialCompletionProvider.cs" />
-    <Compile Include="Completion\CompletionProviders\ReferenceDirectiveCompletionProvider.ItemRules.cs" />
-    <Compile Include="Completion\CompletionProviders\ReferenceDirectiveCompletionProvider.cs" />
     <Compile Include="Completion\CompletionProviders\XmlDocCommentCompletion\XmlDocCommentCompletionItemRules.cs" />
     <Compile Include="Completion\CompletionProviders\XmlDocCommentCompletion\XmlDocCommentCompletionProvider.cs" />
+    <Compile Include="Completion\FileSystem\GlobalAssemblyCacheCompletionHelper.cs" />
+    <Compile Include="Completion\FileSystem\ReferenceDirectiveCompletionProvider.cs" />
+    <Compile Include="Completion\FileSystem\ReferenceDirectiveCompletionProvider.ItemRules.cs" />
     <Compile Include="ContentType\ContentTypeDefinitions.cs" />
     <Compile Include="CSharpEditorResources.Designer.cs">
       <AutoGen>True</AutoGen>
@@ -286,6 +286,7 @@
     <None Include="packages.config" />
     <PublicAPI Include="PublicAPI.txt" />
   </ItemGroup>
+  <ItemGroup />
   <ImportGroup Label="Targets">
     <Import Project="..\..\..\build\Targets\VSL.Imports.targets" />
     <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
