+++ src/EditorFeatures/CSharp/Completion/FileSystem/GlobalAssemblyCacheCompletionHelper.cs
@@ -5,12 +5,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Completion;
-using Microsoft.CodeAnalysis.Completion.Providers;
+using Microsoft.CodeAnalysis.Editor.Implementation.IntelliSense.Completion.FileSystem;
 using Microsoft.CodeAnalysis.Shared.Extensions;
 using Microsoft.CodeAnalysis.Text;
 using Roslyn.Utilities;
 
-namespace Microsoft.CodeAnalysis.Editor.CSharp.Completion.CompletionProviders
+namespace Microsoft.CodeAnalysis.Editor.CSharp.Completion.FileSystem
 {
     internal sealed class GlobalAssemblyCacheCompletionHelper
     {
