+++ src/EditorFeatures/CSharpTest/Completion/ReferenceDirectiveCompletionProviderTests.cs
@@ -4,7 +4,7 @@ using System;
 using System.IO;
 using System.Linq;
 using Microsoft.CodeAnalysis.Completion;
-using Microsoft.CodeAnalysis.Editor.CSharp.Completion.CompletionProviders;
+using Microsoft.CodeAnalysis.Editor.CSharp.Completion.FileSystem;
 using Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.Completion.CompletionProviders;
 using Roslyn.Test.Utilities;
 using Xunit;
