+++ src/EditorFeatures/Core/EditorFeatures.csproj
@@ -413,6 +413,11 @@
     <Compile Include="Implementation\GoToDefinition\IGoToDefinitionService.cs" />
     <Compile Include="Implementation\Intellisense\Completion\CompletionProviders\MemberInsertingCompletionItemRules.cs" />
     <Compile Include="Implementation\Intellisense\Completion\CompletionProviders\XmlDocCommentCompletion\AbstractXmlDocCommentCompletionItemRules.cs" />
+    <Compile Include="Implementation\Intellisense\Completion\FileSystem\CurrentWorkingDirectoryDiscoveryService.cs" />
+    <Compile Include="Implementation\Intellisense\Completion\FileSystem\FileSystemCompletionHelper.cs" />
+    <Compile Include="Implementation\Intellisense\Completion\FileSystem\ICurrentWorkingDirectoryDiscoveryService.cs" />
+    <Compile Include="Implementation\Intellisense\Completion\FileSystem\IOUtilities.cs" />
+    <Compile Include="Implementation\Intellisense\Completion\FileSystem\PathCompletionUtilities.cs" />
     <Compile Include="Implementation\KeywordHighlighting\AbstractKeywordHighlighter.cs" />
     <Compile Include="Implementation\KeywordHighlighting\HighlighterTagProducer.cs" />
     <Compile Include="Implementation\KeywordHighlighting\HighlighterViewTaggerProvider.cs" />
@@ -501,7 +506,6 @@
     <Compile Include="Implementation\Intellisense\Completion\Controller_TypeChar.cs" />
     <Compile Include="Implementation\Intellisense\Completion\DescriptionModifyingCompletionItem.cs" />
     <Compile Include="Implementation\Intellisense\Completion\Model.cs" />
-    <Compile Include="Implementation\Intellisense\Completion\PathCompletionUtilities.cs" />
     <Compile Include="Implementation\Intellisense\Completion\Presentation\CompletionPresenter.cs" />
     <Compile Include="Implementation\Intellisense\Completion\Presentation\CompletionPresenterSession.cs" />
     <Compile Include="Implementation\Intellisense\Completion\Presentation\CompletionSet2.cs" />
@@ -834,6 +838,7 @@
     </None>
     <PublicAPI Include="PublicAPI.txt" />
   </ItemGroup>
+  <ItemGroup />
   <ImportGroup Label="Targets">
     <Import Project="..\..\..\build\Targets\VSL.Imports.targets" />
     <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
