+++ src/EditorFeatures/Core/Implementation/IntelliSense/Completion/FileSystem/CurrentWorkingDirectoryDiscoveryService.cs
@@ -1,9 +1,9 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System.IO;
-using Microsoft.CodeAnalysis.Text;
+using Microsoft.VisualStudio.Text;
 
-namespace Microsoft.CodeAnalysis.Completion.Providers
+namespace Microsoft.CodeAnalysis.Editor.Implementation.IntelliSense.Completion.FileSystem
 {
     internal class CurrentWorkingDirectoryDiscoveryService : ICurrentWorkingDirectoryDiscoveryService
     {
@@ -16,6 +16,14 @@ namespace Microsoft.CodeAnalysis.Completion.Providers
         public string CurrentDirectory
         {
             get { return Directory.GetCurrentDirectory(); }
+        }
+
+        public static ICurrentWorkingDirectoryDiscoveryService GetService(ITextSnapshot textSnapshot)
+        {
+            ICurrentWorkingDirectoryDiscoveryService result;
+            return textSnapshot.TextBuffer.Properties.TryGetProperty(typeof(ICurrentWorkingDirectoryDiscoveryService), out result)
+                ? result
+                : Instance;
         }
     }
 }
