+++ src/EditorFeatures/Core/Implementation/IntelliSense/Completion/FileSystem/ICurrentWorkingDirectoryDiscoveryService.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-namespace Microsoft.CodeAnalysis.Completion.Providers
+namespace Microsoft.CodeAnalysis.Editor.Implementation.IntelliSense.Completion.FileSystem
 {
     internal interface ICurrentWorkingDirectoryDiscoveryService
     {
